refactor: identify public and internal types and interfaces (#83)
BREAKING CHANGE: Change signature from onBodyPage({ offset }: any): void to onBodyPage(offset: number): void
