+++ projects/ngx-datatable/src/lib/components/body/body-group-header.directive.ts
@@ -1,6 +1,6 @@
 import { ContentChild, Directive, EventEmitter, Input, Output, TemplateRef } from '@angular/core';
 import { DatatableGroupHeaderTemplateDirective } from './body-group-header-template.directive';
-import { GroupContext } from '../../types/cell-context.type';
+import { GroupContext } from '../../types/public.types';
 
 @Directive({ selector: 'ngx-datatable-group-header' })
 export class DatatableGroupHeaderDirective<TRow = any> {
