+++ projects/ngx-datatable/src/lib/components/body/selection.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
-import { SelectionType } from '../../types/selection.type';
 import { selectRows, selectRowsBetween } from '../../utils/selection';
 import { Keys } from '../../utils/keys';
+import { SelectionType } from '../../types/public.types';
 
 export interface Model<TRow> {
   type: string;
