+++ projects/ngx-datatable/src/lib/components/columns/column-header.directive.ts
@@ -1,5 +1,5 @@
 import { Directive } from '@angular/core';
-import { HeaderCellContext } from '../../types/cell-context.type';
+import { HeaderCellContext } from '../../types/public.types';
 
 @Directive({ selector: '[ngx-datatable-header-template]' })
 export class DataTableColumnHeaderDirective {
