+++ projects/ngx-datatable/src/lib/components/footer/footer.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 import { DatatableFooterDirective } from './footer.directive';
-import { PagerPageEvent } from '../../types/page-event.type';
+import { PagerPageEvent } from '../../types/public.types';
 @Component({
   selector: 'datatable-footer',
   template: `
