+++ projects/ngx-datatable/src/lib/components/footer/footer.directive.ts
@@ -1,6 +1,6 @@
 import { ContentChild, Directive, Input, TemplateRef } from '@angular/core';
 import { DataTableFooterTemplateDirective } from './footer-template.directive';
-import { FooterContext } from '../../types/footer-context';
+import { FooterContext } from '../../types/public.types';
 
 @Directive({ selector: 'ngx-datatable-footer' })
 export class DatatableFooterDirective {
