--- projects/ngx-datatable/src/lib/types/click.type.ts
@@ -1,4 +1,0 @@
-export enum ClickType {
-  single = 'single',
-  double = 'double'
-}
