--- projects/ngx-datatable/src/lib/types/column-group-width.type.ts
@@ -1,6 +1,0 @@
-export interface ColumnGroupWidth {
-  left: number;
-  center: number;
-  right: number;
-  total: number;
-}
