--- projects/ngx-datatable/src/lib/types/column-mode.type.ts
@@ -1,5 +1,0 @@
-export enum ColumnMode {
-  standard = 'standard',
-  flex = 'flex',
-  force = 'force'
-}
