--- projects/ngx-datatable/src/lib/types/column-pin.type.ts
@@ -1,8 +1,0 @@
-import { TableColumn } from './table-column.type';
-
-export type PinDirection = 'left' | 'center' | 'right';
-
-export interface PinnedColumns {
-  type: PinDirection;
-  columns: TableColumn[];
-}
