--- projects/ngx-datatable/src/lib/types/contextmenu.type.ts
@@ -1,4 +1,0 @@
-export enum ContextmenuType {
-  header = 'header',
-  body = 'body'
-}
