--- projects/ngx-datatable/src/lib/types/detail-context.type.ts
@@ -1,9 +1,0 @@
-import { Observable } from 'rxjs';
-import { RowOrGroup } from './group.type';
-
-export interface RowDetailContext<TRow> {
-  row: RowOrGroup<TRow>;
-  expanded: boolean;
-  rowIndex: number;
-  disableRow$?: Observable<boolean>;
-}
