--- projects/ngx-datatable/src/lib/types/footer-context.ts
@@ -1,7 +1,0 @@
-export interface FooterContext {
-  rowCount: number;
-  pageSize: number;
-  selectedCount: number;
-  curPage: number;
-  offset: number;
-}
