--- projects/ngx-datatable/src/lib/types/group.type.ts
@@ -1,10 +1,0 @@
-/** A Group row */
-export interface Group<TRow> {
-  /** The value by which to rows are grouped. */
-  key: TRow[keyof TRow];
-  /** All rows that are part of the group. */
-  value: TRow[];
-}
-
-/** Type for either a row or a group */
-export type RowOrGroup<TRow> = TRow | Group<TRow>;
