--- projects/ngx-datatable/src/lib/types/page-event.type.ts
@@ -1,14 +1,0 @@
-export interface PageEvent {
-  count: number;
-  pageSize: number;
-  limit: number;
-  offset: number;
-}
-
-export interface BodyPageEvent {
-  offset: number;
-}
-
-export interface PagerPageEvent {
-  page: number;
-}
