--- projects/ngx-datatable/src/lib/types/pages.type.ts
@@ -1,4 +1,0 @@
-export interface Page {
-  number: number;
-  text: string;
-}
