--- projects/ngx-datatable/src/lib/types/resize.type.ts
@@ -1,7 +1,0 @@
-import { TableColumn } from './table-column.type';
-
-export interface ColumnResizeEvent {
-  column: TableColumn;
-  prevValue: number;
-  newValue: number;
-}
