--- projects/ngx-datatable/src/lib/types/scroll.type.ts
@@ -1,4 +1,0 @@
-export interface ScrollEvent {
-  offsetY: number;
-  offsetX: number;
-}
