--- projects/ngx-datatable/src/lib/types/selection.type.ts
@@ -1,7 +1,0 @@
-export enum SelectionType {
-  single = 'single',
-  multi = 'multi',
-  multiClick = 'multiClick',
-  cell = 'cell',
-  checkbox = 'checkbox'
-}
