--- projects/ngx-datatable/src/lib/types/sort-prop-dir.type.ts
@@ -1,7 +1,0 @@
-import { SortDirection } from './sort-direction.type';
-import { TableColumnProp } from './table-column.type';
-
-export interface SortPropDir {
-  dir: SortDirection | 'desc' | 'asc';
-  prop: TableColumnProp;
-}
