--- projects/ngx-datatable/src/lib/types/sort.type.ts
@@ -1,4 +1,0 @@
-export enum SortType {
-  single = 'single',
-  multi = 'multi'
-}
