+++ projects/ngx-datatable/src/lib/utils/sort.ts
@@ -1,7 +1,5 @@
 import { getterForProp } from './column-prop-getters';
-import { SortType } from '../types/sort.type';
-import { SortDirection } from '../types/sort-direction.type';
-import { SortPropDir } from '../types/sort-prop-dir.type';
+import { SortDirection, SortPropDir, SortType } from '../types/public.types';
 /**
  * Gets the next sort direction
  */
